% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{AlignPlots}
\alias{AlignPlots}
\title{Align ggplo2 plots vertically}
\usage{
AlignPlots(..., plots = NULL, heights = NULL)
}
\arguments{
\item{...}{ggplo2 plots}

\item{plots}{list with ggplot2 plots, default to NULL}

\item{heights}{the heights of the single plots. If NULL an equal height of 1 is asusmed for all plots. Default to NULL}
}
\value{
arranged plots
}
\description{
Align ggplo2 plots vertically
}

