% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{CIndexTest}
\alias{CIndexTest}
\title{CIndexTest}
\usage{
CIndexTest(formula, var, data, N = 10000, seed = 123, cores = 2, ...)
}
\arguments{
\item{formula}{model formula with a Surv object on the left side}

\item{var}{the name of variable to test with C-Index. Will be added to the formula if not included}

\item{data}{the data frame for model fitting. Must contain all variables in the formula}

\item{N}{number of permutations for the test, default to 10,000}

\item{cores}{number of cores to use for permutation}
}
\value{
list with the original C-Index, the permutation C-Indices and the p-value (one sided, alternative="higher")
}
\description{
Tests for a variable the Improvement in C-Index of a Cox model
}

