% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{CoxP}
\alias{CoxP}
\title{Helper function to calculate Cox Model p values}
\usage{
CoxP(x)
}
\arguments{
\item{x}{a coxph objejct}
}
\value{
the p-values from the cox model
}
\description{
Helper function to calculate Cox Model p values
}
\note{
code is from the print.coxph function in the survival package by Terry Therneau
}

