% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{CreateSurvivalFrame}
\alias{CreateSurvivalFrame}
\title{Creates a survival data frame}
\usage{
CreateSurvivalFrame(f.survfit)
}
\arguments{
\item{s.survfit}{a survival fit object}
}
\value{
a data frame with survival data
}
\description{
The functions creates an data frame needed for \link{qplot_survival}
}
\author{
R. Saccilotto
}

