% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{ExtractHR}
\alias{ExtractHR}
\title{'coxph' and related objects}
\usage{
ExtractHR(object, level = 0.95, p.level = NULL, ...)
}
\arguments{
\item{object}{a 'coxph' object}

\item{level}{the confidence level (has to be between 0 and 1). Default to 0.95}

\item{p.level}{a number between 0 and 1 that will be used as lower bound to display p-values. P-values below this level are just printed as "<p.level". If it is NULL, p-value are rounded to three decimals. Default to NULL.}

\item{...}{further arguments to \code{confint}}
}
\value{
matrix with hazard-ratios (exponential of coefficients), confidence intervals, and p values
}
\description{
'coxph' and related objects
}

