% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{ExtractOR}
\alias{ExtractOR}
\title{Function to extract odds-ratios, confidence intervals, and p values as given by the 'summary' function for
'glm' and related objects}
\usage{
ExtractOR(object, level = 0.95, p.level = NULL, ci.type = c("pl", "wald"),
  ...)
}
\arguments{
\item{object}{a 'glm' object}

\item{level}{the confidence level (has to be between 0 and 1). Default to 0.95}

\item{p.level}{a number between 0 and 1 that will be used as lower bound to display p-values. P-values below this level are just printed as "<p.level". If it is NULL, p-value are rounded to three decimals. Default to NULL.}

\item{ci.type}{the type of the confidence intervals. Use 'pl' for profile-likelihood based confidence intervals and 'wald' for Wald based intervals. Default is 'pl'. If profile-likelihood based cannot be computed, wald based intervals will be used as fallback.}

\item{...}{further arguments to \code{confint}}
}
\value{
matrix with odds-ratios (exponential of coefficients), confidence intervals, and p values
}
\description{
Function to extract odds-ratios, confidence intervals, and p values as given by the 'summary' function for
'glm' and related objects
}

