% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggkm}
\alias{ggkm}
\title{Kaplan Meier Plot from Matt Cooper (http://mcfromnz.wordpress.com/2012/05/05/kaplan-meier-survival-plot-with-at-risk-table-by-sub-groups/)}
\usage{
ggkm(sfit, sdiff = NULL, table = TRUE, returns = FALSE, plot = T,
  color = TRUE, xlabs = "Time", ylabs = "Survival Probability",
  xlims = c(0, max(sfit$time)), ylims = c(0, 1), ystratalabs = NULL,
  ystrataname = NULL, timeby = 100, main = "Kaplan-Meier Plot",
  pval = TRUE, pval.size = 4, probs = NULL, subs = NULL, ...)
}
\arguments{
\item{sfit}{a \code{\link[survival]{survfit}} object}

\item{sdiff}{a \code{\link[survival]{survdiff}} object, used for compuation of exact p-values. If no survdiff object is given computation of p-values is based on the survfit object}

\item{table}{logical: Create a table graphic below the K-M plot, indicating at-risk numbers?}

\item{returns}{logical: if \code{TRUE}, return an arrangeGrob object}

\item{plot}{logical: if \code{TRUE}, plot the graph}

\item{color}{logical: if \code{TRUE}, use a colored scheme, otherwise black & white}

\item{xlabs}{x-axis label}

\item{ylabs}{y-axis label}

\item{xlims}{the limit of x axis}

\item{ylims}{the limit on the y axis}

\item{ystratalabs}{The strata labels. \code{Default = levels(summary(sfit)$strata)}}

\item{ystrataname}{The legend name. Default = "Strata"}

\item{timeby}{numeric: control the granularity along the time-axis}

\item{main}{plot title}

\item{pval}{logical: add the pvalue to the plot?}

\item{pval.size}{size of the pvalue in the plot. Default to 3}

\item{probs}{numeric: vector with survail probabilities that will be marked by dotted lines, dfault to NULL}

\item{subs}{default to NULL}
}
\description{
Kaplan Meier Plot from Matt Cooper (http://mcfromnz.wordpress.com/2012/05/05/kaplan-meier-survival-plot-with-at-risk-table-by-sub-groups/)
}
\author{
Matt Cooper
\url{http://statbandit.wordpress.com/2011/03/08/an-enhanced-kaplan-meier-plot/}
}

