% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{LogRankP}
\alias{LogRankP}
\title{Helper function to calculate Logrank Test p values}
\usage{
LogRankP(x)
}
\arguments{
\item{x}{a survfit objejct}
}
\value{
the logrank test p-value
}
\description{
Helper function to calculate Logrank Test p values
}
\note{
code is from the print.survdiff function in the survival package by Terry Therneau
}

