% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{NormalizeQuantilesTemplate}
\alias{NormalizeQuantilesTemplate}
\title{Normalize Quantiles with another data matrix as template to get the quantiles}
\usage{
NormalizeQuantilesTemplate(x, template)
}
\arguments{
\item{x:}{the data matrix or vector we want to normalize}

\item{template:}{a data matrix or vector providing the quantiles}
}
\value{
the data matrix/vector normalized to the same quantiles like the template
}
\description{
Normalize Quantiles with another data matrix as template to get the quantiles
}

