% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{ParseMatrixFile}
\alias{ParseMatrixFile}
\title{Helper function to read clinical annotation from GEO Matrix Series files}
\usage{
ParseMatrixFile(file, start.line, end.line, pattern.search,
  pattern.extract = "^.*:\\\\s*", drop.first.col = T, ...)
}
\arguments{
\item{file}{the name of the matrix series file}

\item{start.line}{the number of the line holding the names of the samples}

\item{end.line}{the line after (!) the last line to read}

\item{pattern.search}{a named character vector giving the search pattern for the different clinical parameters, names should be the names of the pararmeters}

\item{pattern.extract}{character vector holding the pattern for extraction the desired parameter value, set to "" if no extraction is need, vector is recycled to same length as  pattern.search}

\item{drop.first.col}{boolean, indicates if the first column of the file contains line descriptors and thus should be dropped, default to TRUE}

\item{...}{further arguments to sub}
}
\value{
a data frame with the parameters specified in pattern.search
}
\description{
Helper function to read clinical annotation from GEO Matrix Series files
}

