% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{print.citest}
\alias{print.citest}
\title{Print function for 'citest' class}
\usage{
\method{print}{citest}(x)
}
\arguments{
\item{x}{object of class 'citest'}
}
\value{
x (invisible)
}
\description{
Print function for 'citest' class
}

