% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{print.ggkm}
\alias{print.ggkm}
\title{Print function for ggkm plots}
\usage{
\method{print}{ggkm}(x)
}
\arguments{
\item{x}{object of class 'ggkm'}
}
\description{
A helper function for seamless call of print to results of a ggkm function
}

