% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{PrintHR}
\alias{PrintHR}
\title{Prints output of Cox model}
\usage{
PrintHR(object, ...)
}
\arguments{
\item{object}{a coxph model}

\item{...}{further argument to texreg}
}
\description{
The function is wrapper around the \link{texreg} function of the \link{texreg} package which replaces the names of coefficients
}

