% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{PrintLatex}
\alias{PrintLatex}
\title{Helper function to make Hmisc's latex easier}
\usage{
PrintLatex(x, file = "", booktabs = T, ctable = F, caption = NULL,
  caption.loc = "top", rowlabel = "", n.cgroup = NULL, cgroup = NULL,
  n.rgroup = NULL, rgroup = NULL, p.columns = NULL, p.level = 0.05,
  p.cmd = "bfseries", cellTexCmds = NULL, sanitize = function(xx) {     xx
  }, ...)
}
\arguments{
\item{file}{the file to be written, default to "" which means no file but stdout}

\item{booktabs}{boolean, whether to use booktabs package or not, default to TRUE}

\item{ctable}{boolean, whether to use ctable package or not, default to FALSE}

\item{caption}{see Hmisc::latex, default to NULL.}

\item{caption.loc}{see Hmisc::latex, default to "top".}

\item{rowlabel}{label for first column with rownames, default to ""}

\item{n.cgroup}{see Hmisc::latex, default to NULL. If NULL but cgroup is not will be set accordingly}

\item{cgroup}{see Hmisc::latex}

\item{n.rgroup}{see Hmisc::latex, default to NULL. If NULL but rgroup is not will be set accordingly}

\item{rgroup}{see Hmisc::latex}

\item{p.columns}{column numbers of columns with p values. If NULL p values are not highlighted. Default to NULL.}

\item{p.level}{p values below that significance level are highligted for better reading (only if p.columns is given). Default to 0.05}

\item{p.cmd}{character string. The latex command used to highlight p values. Default to "bfseries" for boldface.}

\item{cellTexCmds}{see Hmisc::latex. Default to NULL.}

\item{sanitize}{function to sanitize row- and colnames. Default is the identical function where nothing will be replaced.}

\item{...}{further arguments to Hmisc::latex}

\item{data}{frame or matrix to be printed}
}
\value{
returns (invisible) a list of class \code{latex}, see Hmisc \code{latex} for details
}
\description{
Helper function to make Hmisc's latex easier
}

