% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{qplot_survival}
\alias{qplot_survival}
\title{Survival (KM) plot for ggplot2}
\usage{
qplot_survival(f.frame, f.CI = "default", f.shape = 3)
}
\arguments{
\item{f.frame}{a survival data frame constructed by \link{CreateSurvivalFrame}}

\item{f.shape}{the shape of censoring indicators}

\item{f.ci}{wheter to draw confidence interval: TRUE, FALSE or 'default'. With 'default' it is dependent wheter a strata is given: no with strata, yes without}
}
\description{
Survival (KM) plot for ggplot2
}
\author{
R. Saccilotto
}

