% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{ReplaceCoefNames}
\alias{ReplaceCoefNames}
\title{Helper function to replace factors in coeffecients by factor: level}
\usage{
ReplaceCoefNames(object)
}
\arguments{
\item{object}{model object}
}
\description{
Helper function to replace factors in coeffecients by factor: level
}

