% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{SplitGraphs}
\alias{SplitGraphs}
\title{SplitGraphs
Splits list of ggplot2 plots into several grids od define size and plots them}
\usage{
SplitGraphs(plot.list, nrow = 3, ncol = 2, ...)
}
\arguments{
\item{plot.list}{the list with gglot2 graphs}

\item{nrow,}{the number of rows for one grid, default to 3}

\item{ncol,}{the number of columns of one grid, default to 2}

\item{...}{additional arguments to grid.arrange}
}
\description{

}
\author{
Stephan Gade
}

