% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{SplitVec}
\alias{SplitVec}
\title{function to split a vector in pieces, needed for e.g. cross validation}
\usage{
SplitVec(x, n)
}
\arguments{
\item{x}{the vector to split}

\item{n}{the number of splits}
}
\description{
function to split a vector in pieces, needed for e.g. cross validation
}

