% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{SummarizeProbes}
\alias{SummarizeProbes}
\title{Function to summarize probes of an expression matrix based on common gene identifiers (e.g gene symbols)}
\usage{
SummarizeProbes(eset, mapping, FUN = median, ...)
}
\arguments{
\item{eset}{the data matrix we want to summarize, the rownames have to be identiefiers in the first column of mappings}

\item{mapping}{a character vector with the IDs we want to use for summarization}
}
\value{
the expression matrix with summarized probes
}
\description{
Function to summarize probes of an expression matrix based on common gene identifiers (e.g gene symbols)
}

